more cast updates

+++ src/BaseDisplay.hh
@@ -64,8 +64,8 @@ public:
   inline int getDepth(void) const { return depth; }
   inline unsigned int getScreenNumber(void) const
     { return screen_number; }
-  inline unsigned short getWidth(void) const { return rect.width; }
-  inline unsigned short getHeight(void) const { return rect.height; }
+  inline unsigned int getWidth(void) const { return rect.width; }
+  inline unsigned int getHeight(void) const { return rect.height; }
   inline const std::string& displayString(void) const
   { return display_string; }
 };
